pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "surrealdb_openai"
version = "0"
description = "Example of RAG using SurrealDB and OpenAI"
authors = [
    {name="Cellan Hall", email="cellan.hall@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11, <3.12"
dependencies = [
    "pandas",
    "wget",
    "pandas-stubs",
    "surrealdb-beta @ git+https://github.com/surrealdb/surrealdb.py",
    "tqdm",
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    "python-dotenv",
]

[project.scripts]
surreal-insert = "surrealdb_openai.insert:surreal_insert"
download-data = "surrealdb_openai.download:download_data"

[tool.ruff]
target-version = "py311"
line-length = 80

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ISC001", "COM812"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
]

